+++ src/main/kotlin/next/step/minesweeper/domain/mine/MineCount.kt
@@ -1,7 +1,8 @@
-package next.step.minesweeper.domain
+package next.step.minesweeper.domain.mine
 
 @JvmInline
 value class MineCount(val count: Int) {
+    
     init {
         require(count > MIN_COUNT) { "지뢰 개수는 ${MIN_COUNT}보다 커야합니다." }
     }
